\name{countsBottomly}
\docType{data}
\alias{countsBottomly}
\title{RNA-Seq data of two mice strains.}
\description{The two common mice strains C57BL/6J (B6) and DBA/2J (D2) 
were used for comparing gene expression measures of RNA-Seq and microarrays.
}
\usage{countsBottomly}
\format{A data matrix of 36229 rows (genes) and 21 columns (samples).}
\source{http://bowtie-bio.sourceforge.net/recount/}
\references{
Bottomly, D., Walter, N. A. R., Hunter, J. E., Darakjian, P., Kawane,
S., Buck, K. J., Searles, R. P., Mooney, M., McWeeney, S. K., and
Hitzemann, R. (2011). \emph{Evaluating gene expression in C57BL/6J
and DBA/2J mouse striatum using RNA-Seq and microarrays.}
Plos One, 6(3), e17820.
}
\keyword{datasets}
