\name{countsGilad}
\docType{data}
\alias{countsGilad}
\title{RNA-Seq data of humans, chimpanzees and rhesus macaques.}
\description{Liver RNA samples of three males and three females 
from each of the species human, chimpanzee and rhesus macaques were sequenced.
}
\usage{countsGilad}
\format{A data matrix of 20689 rows (genes) and 18 columns (samples).}
\source{ftp://ftp.ncbi.nlm.nih.gov/pub/geo/DATA/supplementary/series/GSE17274/GSE17274_ReadCountPerLane.txt.gz}
\references{
Blekhman, R., Marioni, J. C., Zumbo, P., Stephens, M., and Gilad,
Y. (2010). \emph{Sex-specific and lineage-specific alternative splicing
in primates.} Genome Res, 20(2), 180-189.
}
\keyword{datasets}
