\name{countsLi}
\docType{data}
\alias{countsLi}
\title{RNA-Seq data of the developmental zones of maize leaves.}
\description{RNA-Sequencing was performed on different locations of the maize plant
leaf. 
}
\usage{countsLi}
\format{A data matrix of 110185 rows (genes) and 12 columns (samples).}
\source{http://www.ncbi.nlm.nih.gov/sra/ accession number:SRP002265}
\references{Li, P., Ponnala, L., Gandotra, N., Wang, L., Si, Y., Tausta, S. L.,
Kebrom, T. H., Provart, N., Patel, R., Myers, C. R., Reidel,
E. J., Turgeon, R., Liu, P., Sun, Q., Nelson, T., and Brutnell,
T. P. (2010). \emph{The developmental dynamics of the maize leaf transcriptome.}
Nat Genet, 42(12), 1060-1067.
}
\keyword{datasets}
