\name{countsMontgomery}
\docType{data}
\alias{countsMontgomery}
\title{RNA-Seq data of 60 European HapMap individuals.}
\description{
The RNA of lymphoblastoid cell lines of 60 HapMap individuals was sequenced
in order to study eQTLs.
}
\usage{countsMontgomery}
\format{A data matrix of 12984 rows (genes) and 60 columns (samples).}
\source{http://bowtie-bio.sourceforge.net/recount/}
\references{Montgomery, S. B., Sammeth, M., Gutierrez-Arcelus, M., Lach,
R. P., Ingle, C., Nisbett, J., Guigo, R., and Dermitzakis, E. T.
(2010). \emph{Transcriptome genetics using second generation sequenc-
ing in a caucasian population.} Nature, 464(7289), 773-777.
}
\keyword{datasets}
