\name{countsPickrell}
\docType{data}
\alias{countsPickrell}
\title{RNA-Seq data of 69 Nigerian HapMap individuals.}
\description{
The RNA of lymphoblastoid cell lines of 69 HapMap individuals was sequenced
in order to study eQTLs.
}
\usage{countsPickrell}
\format{A data matrix of 12984 rows (genes) and 69 columns (samples).}
\source{http://bowtie-bio.sourceforge.net/recount/}
\references{Pickrell, J. K., Marioni, J. C., Pai, A. A., Degner, J. F., Engelhardt,
B. E., Nkadori, E., Veyrieras, J.-B., Stephens, M., Gilad, Y., and
Pritchard, J. K. (2010). \emph{Understanding mechanisms underlying
human gene expression variation with RNA sequencing.} Nature,
464(7289), 768-772.
}
\keyword{datasets}
